PR Review comments
* Fix WarningsProvider typo and move to a separate module
* Add documentation to LoadConfig
* Add test for emitting config warnings
